multi_agent_supervisor/requirements.txt
langgraph
langchain
langchain-google-genai
langchain-community
tavily-python
python-dotenv
defusedxml
langchain-experimental
matplotlib
